requirements.txt
PyYAML
psycopg2-binary
httpx
